# model/user.go
package model

import "gorm.io/gorm"

type User struct {
	gorm.Model
	Username string `json:"username" gorm:"unique"`
	Email    string `json:"email"    gorm:"unique"`
	Password string `json:"password"`
}